operator/requirements.txt
einops==0.6.1
fake_bpy_module_latest==20230526
h5py==3.8.0
imageio==2.30.0
matplotlib==3.7.1
numpy==1.24.3
omegaconf==2.3.0
opencv_python==4.7.0.72
Pillow==9.0.1
PyMCubes==0.1.4
pytorch3d==0.7.4
scipy==1.10.1
tensorboard==2.6.0
termcolor==2.3.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
trimesh==3.21.7